requirements.txt
beautifulsoup4==4.14.3
fastapi==0.135.2
neo4j==6.1.0
openai==2.30.0
pandas==2.0.3
pydantic==2.12.5
python-dotenv==1.2.2
PyYAML==6.0.3
Requests==2.33.0
selenium==4.41.0
streamlit==1.28.0
uvicorn==0.42.0
webdriver_manager==4.0.2
